+++ R/screen.R
@@ -8,11 +8,8 @@ print_screen <- function(ht, ...) cat(to_screen(ht, ...))
 #'
 #' @param ht A huxtable.
 #' @param ... Passed on to \code{to_screen}.
-#' @param borders Print 'horizontal' borders, 'vertical' borders, 'both' or 'neither'. May be abbreviated.
-#' @param blank   Character to print for cell divisions with no border. If \code{NULL}, print a space ' ' but don't print
-#'   empty horizontal borders.
+#' @param blank   Character to print for cell divisions with no border.
 #' @param colnames Whether or not to print colum names.
-#' @param colnames_color Color to print column names. Note: this won't work in RStudio.
 #'
 #' @return \code{to_screen} returns a string. \code{print_screen} prints the string and returns \code{NULL}.
 #'
@@ -28,184 +25,46 @@ print_screen <- function(ht, ...) cat(to_screen(ht, ...))
 #' align(ht)[1,] <- 'left'
 #' print_screen(ht)
 #' print_screen(ht, blank = '.')
-#' print_screen(ht, borders = 'horiz')
 to_screen  <- function (ht, ...) UseMethod('to_screen')
-
-#
-# to_screen2 <- function (ht, borders = c('both', 'horizontal', 'vertical', 'neither'), blank = NULL, colnames = TRUE, colnames_color = 'blue', max_width = 80, ...) {
-#   borders <- match.arg(borders)
-#   collapse_horiz <- is.null(blank)
-#   if (is.null(blank)) blank <- ' '
-#   if (colnames) ht <- add_colnames(ht)
-#   my_colour <- if (is.null(colnames_color)) identity else crayon::make_style(colnames_color)
-#
-#   dc               <- display_cells(ht, all = FALSE)
-#   drow_mat         <- as.matrix(dc[,c('display_row', 'display_col')])
-#   contents         <- clean_contents(ht, type = 'screen')
-#   dc$contents      <- contents[drow_mat]
-#   dc$wrap          <- wrap(ht)[drow_mat]
-#   dc$top_border    <- top_border(ht)[drow_mat]
-#   dc$bottom_border <- bottom_border(ht)[drow_mat]
-#   dc$left_border   <- left_border(ht)[drow_mat]
-#   dc$right_border  <- right_border(ht)[drow_mat]
-#   dc <- dc[order(dc$colspan),]
-#
-#   # this represents character widths on screen for each column
-#   # -3 because have to account for borders
-#   # pattern is:
-#   # BSCCCCCSBSCCCCSBSCCCSB
-#   # where B is border S is space C is cell content
-#   scr_w <- rep(floor(max_width/ncol(ht)) - 3, ncol(ht))
-#   scr_w <- pmax(0, scr_w)
-#   # allocate screen colwidths as follows:
-#   # for each column, find cells overlapping that column with colspan n
-#   # take the max of nowrap cells
-#   # if the sum of widths of the cols, + 3 per internal border, is less, increase it
-#   # if we hit max_width, exit the loop
-#   dcnw <- dc[! dc$wrap,]
-#   for (j in seq_len(nrow(dcnw))) {
-#     dcr <- dcnw[j,]
-#     cell_width <- nchar(dcr$contents, type = 'width')
-#     col_width <- sum(scr_w[dcr$display_col:dcr$end_col]) + 3 * (dcr$colspan - 1)
-#     if ((excess <- cell_width - colwidth) > 0) {
-#       scr_w[dcr$display_col:dcr$end_col] <- scr_w[dcr$display_col:dcr$end_col] +
-#             ceiling(excess/dcr$colspan)
-#       # 2 border cols take 2 chars, internal ones take 3 chars
-#       total_width <- sum(scr_w) + 4 + 3 * (ncol(dcr) - 1)
-#       if ((excess2 <- total_width - max_width) > 0) {
-#         scr_w[dcr$display_col:dcr$end_col] <- scr_w[dcr$display_col:dcr$end_col] -
-#               ceiling(excess2/dcr$colspan)
-#       }
-#     }
-#   }
-#
-#   char_widths <- interleave(scr_w, c(2, rep(3, length(scr_w) - 1), 2))
-#   for (row in 0:nrow(ht)) {
-#     dcr  <- dc[dc$display_row == row, ]
-#     dcr2 <- dc[dc$display_row == row + 1, ]
-#     border_segs <- rep(FALSE, ncol(ht))
-#     corner_segs <- rep(FALSE, ncol(ht) + 1)
-#     side_segs   <- rep(FALSE, ncol(ht) + 1)
-#     for (j in seq_len(nrow(dcr))) {
-#       if (dcr$bottom_border[j] > 0) border_segs[dcr$display_col[j]:dcr$end_col[j]] <- TRUE
-#       if (dcr$left_border[j] > 0) corner_segs[dcr$display_col[j]] <- TRUE
-#       if (dcr$right_border[j] > 0) corner_segs[dcr$end_col[j]] <- TRUE
-#     }
-#     for (j in seq_along(nrow(dcr2))){
-#       if (dcr2$top_border[j] > 0) border_segs[dcr2$display_col[j]:dcr2$end_col[j]] <- TRUE
-#       if (dcr2$left_border[j] > 0) {
-#         corner_segs[dcr2$display_col[j]] <- TRUE
-#         side_segs[dcr2$display_col[j]] <- TRUE
-#       }
-#       if (dcr2$right_border[j] > 0) {
-#         corner_segs[dcr2$end_col[j]] <- TRUE
-#         side_segs[dcr2$end_col[j]] <- TRUE
-#       }
-#     }
-#
-#
-#   }
-#
-#
-#
-#   # now, calculate each screen row:
-#   # border rows are easy
-#   # cell rows: nowrap cells are truncated to sum of widths + 3 per int border
-#   #            wrap cells are wrapped, and this creates a new screen row
-#   #            horiz borders are superimposed - we know where they are
-#
-#
-# }
 
 
 #' @export
 #' @rdname to_screen
-to_screen.huxtable <- function (ht, borders = c('both', 'horizontal', 'vertical', 'neither'), blank = NULL,
-      colnames = TRUE, colnames_color = 'blue', ...) {
-  borders <- match.arg(borders)
-  collapse_horiz <- is.null(blank)
-  if (is.null(blank)) blank <- ' '
-  if (colnames) {
-    ht <- add_colnames(ht)
-  }
-  my_colour <- if (is.null(colnames_color)) identity else crayon::make_style(colnames_color)
+to_screen.huxtable <- function (ht, blank = ' ', colnames = TRUE, ...) {
+  charmat_data <- character_matrix(ht, inner_border_h = 3, outer_border_h = 2, inner_border_v = 1, outer_border_v = 1)
+  charmat <- charmat_data$charmat
+  border_rows <- charmat_data$border_rows
+  border_cols <- charmat_data$border_cols
+  border_cols[-1] <- border_cols[-1] + 1 # middle of 3 for interior, last of 2 for last outer
 
   dc <- display_cells(ht, all = FALSE)
-  drow_mat <- as.matrix(dc[, c('display_row', 'display_col')])
-  contents <- clean_contents(ht, type = 'screen')
-  dc$contents <- contents[drow_mat]
-
-  dc <- dc[order(dc$colspan), ]
-  border_chars   <- 3
-
-  dc$widths <- nchar(dc$contents, type = 'width')
-
-  # widths of actual columns, not including borders
-  max_widths <- rep(0, ncol(ht))
-  for (r in 1:nrow(dc)) {
-    width <- dc$widths[r]
-    cols <- with(dc[r, ], display_col:end_col)
-    if (sum(max_widths[cols]) + border_chars * (dc$colspan[r] - 1) < width) {
-      max_widths[cols] <- pmax(max_widths[cols], ceiling(width / dc$colspan[r]))
-    }
+  for (r in seq_len(nrow(dc))) {
+    dcell <- dc[r, ]
+    drow <- dcell$display_row
+    dcol <- dcell$display_col
+    end_row <- dcell$end_row + 1 # carry on to next row/col
+    end_col <- dcell$end_col + 1
+    bdrs <- get_all_borders(ht, drow, dcol)
+    charmat[ border_rows[drow]:border_rows[end_row], border_cols[dcol] ]     <- blank
+    charmat[ border_rows[drow]:border_rows[end_row], border_cols[end_col] ]  <- blank
+    charmat[ border_rows[drow], border_cols[dcol]:border_cols[end_col] ]     <- blank
+    charmat[ border_rows[end_row], border_cols[dcol]:border_cols[end_col] ]  <- blank
+    if (bdrs$left > 0)   charmat[ border_rows[drow]:border_rows[end_row], border_cols[dcol] ]     <- '|'
+    if (bdrs$right > 0)  charmat[ border_rows[drow]:border_rows[end_row], border_cols[end_col] ]  <- '|'
+    if (bdrs$top > 0)    charmat[ border_rows[drow], border_cols[dcol]:border_cols[end_col] ]     <- '-'
+    if (bdrs$bottom > 0) charmat[ border_rows[end_row], border_cols[dcol]:border_cols[end_col] ]  <- '-'
   }
 
-  charmat         <- matrix(' ', 2 * nrow(ht) + 1, sum(max_widths) + (ncol(ht) + 1) * border_chars)
-  border_cells    <- matrix(0, nrow(charmat), ncol(charmat)) # 0 = not a border; 1 = no border; > 1 = border
-  corner_cells    <- matrix(FALSE, nrow(charmat), ncol(charmat))
-  for (r in 1:nrow(dc)) {
-    dcr     <- dc[r, ]
-    drow    <- dcr$display_row
-    dcol    <- dcr$display_col
-    end_col <- dcol + dcr$colspan - 1
-    end_row <- drow + dcr$rowspan - 1
-
-    # content
-    chars       <- strsplit(dcr$contents, '')[[1]]
-    char_row    <- drow * 2
-    total_width <- sum(max_widths[dcol:end_col]) + border_chars * (dcr$colspan - 1) # include internal borders
-    start_char  <- sum(max_widths[seq_len(dcol - 1)]) + border_chars * dcol
-    while (length(chars) > 0) {
-      idx      <- 1:min(total_width, length(chars))
-      space    <- max(total_width - length(chars), 0)
-      char_idx <- switch(align(ht)[drow, dcol],
-              left  = start_char + idx,
-              right = start_char + space + idx,
-              center = start_char + floor(space / 2) + idx
-            )
-      charmat[char_row, char_idx ] <- chars[idx]
-      chars <- chars[-idx]
-      char_row <- char_row + 1
-    }
-
-    bdr_idx_cols <- (start_char - 1):(start_char + total_width + 2)
-    bord <- get_all_borders(ht, drow, dcol)
-    border_cells[drow * 2 - 1, bdr_idx_cols] <- pmax(border_cells[drow * 2 - 1, bdr_idx_cols], 1 + bord$top)
-    border_cells[end_row * 2 + 1, bdr_idx_cols] <- pmax(border_cells[end_row * 2 + 1, bdr_idx_cols],
-            1 + bord$bottom)
-    border_cells[(drow * 2):(end_row * 2), start_char - 1] <-
-            pmax(border_cells[(drow * 2):(end_row * 2), start_char - 1], 1 + bord$left)
-    border_cells[(drow * 2):(end_row * 2), start_char + total_width + 2] <-
-          pmax(border_cells[(drow * 2):(end_row * 2), start_char + total_width + 2], 1 + bord$right)
-    corner_cells[drow * 2 + c(-1, 1), c(start_char - 1, start_char + total_width + 2)] <- TRUE
-  }
-
-  charmat[border_cells > 0] <- blank
-  if (borders %in% c('both', 'horizontal')) charmat[border_cells > 1 & row(charmat) %% 2]   <- '-'
-  if (borders %in% c('both', 'vertical'))   charmat[border_cells > 1 & ! row(charmat) %% 2] <- '|'
-  if (borders == 'vertical') charmat[border_cells > 1 & row(charmat) %% 2 & corner_cells] <- '|'
-  if (collapse_horiz) {
-    empty_borders <- apply(border_cells, 1, function (x) all(x[ -c(1, ncol(charmat)) ] == 1))
-    charmat <- charmat[ !empty_borders, ]
-  }
-  result <- apply(charmat, 1, paste0, collapse = '')
-
-  if (colnames) result[2 - collapse_horiz] <- my_colour(result[2 - collapse_horiz])
-  result <- paste0(result, collapse = '\n')
+  result <- paste((apply(charmat, 1, paste0, collapse='')), collapse='\n')
   if (! is.na(cap <- caption(ht))) {
+    poss_pos <- c('left', 'center', 'right')
+    hpos <- if (any(found <- sapply(poss_pos, grepl, x = caption_pos(ht)))) poss_pos[found] else position(ht)
+    cap <- str_pad(cap, hpos, ncol(charmat))
     result <- if (grepl('top', caption_pos(ht))) paste0(cap, '\n', result) else paste0(result, '\n', cap)
   }
-  result <- paste0(result, '\n')
+  if (colnames) {
+    result <- paste0(result, '\n\n', 'Column names: ', paste(colnames(ht), collapse = ', '))
+  }
 
   result
 }
@@ -220,6 +79,7 @@ print_md <- function(ht, ...) cat(to_md(ht, ...))
 #'
 #' @param ht        A huxtable.
 #' @param max_width Max width in on-screen characters of the result.
+#' @param header    Logical. Print the first row as a header?
 #' @param ...       Arguments passed to methods.
 #'
 #' @return \code{to_md} returns a string. \code{print_md} prints the string and returns
@@ -237,62 +97,138 @@ print_md <- function(ht, ...) cat(to_md(ht, ...))
 #' print_md(ht)
 to_md <- function(ht, ...) UseMethod('to_md')
 
+
 #' @export
 #' @rdname to_md
-to_md.huxtable <- function(ht, max_width = 80, ...) {
-  cw <- col_width(ht)
-  if (! is.numeric(cw) || anyNA(cw)) cw <- rep(1, ncol(ht))
-  cw <- floor(cw / sum(cw) * (max_width - ncol(ht) + 1))
-  width <- sum(cw) + ncol(ht) - 1
-  ncharw <- function(x) nchar(x, type = 'width')
-  dcells <- display_cells(ht, all = FALSE)
-  if (any(dcells$shadowed)) warning("Markdown cannot handle cells with colspan/rowspan > 1")
-  dcells <- dcells[! dcells$shadowed, ]
-  result <- str_rep('-', width)
-  result <- paste0(result, '\n')
-  contents <- clean_contents(ht, type = 'markdown')
-  dcells$contents <- contents[as.matrix(dcells[, c('display_row', 'display_col')])]
+to_md.huxtable <- function(ht, max_width = 80, header = TRUE, ...) {
+  if (any(colspan(ht) > 1 | rowspan(ht) > 1)) warning("Markdown cannot handle cells with colspan/rowspan > 1")
 
   align <- align(ht)
   if (any(apply(align, 2, function(x) length(unique(x)) > 1)))
-        warning("Can't vary column alignment in markdown; using first row")
-  align <- align[1, ]
+    warning("Can't vary column alignment in markdown; using first row")
+  ht <- set_align(ht, align[1, ], byrow = TRUE)
 
-  # for every row:
-  #   print the first cw characters of display cells in that row
-  #   if there are any left over, start a new row
-  for (myrow in 1:nrow(ht)) {
-    row_chars <- rep('', ncol(ht))
-    dcr <- dcells[dcells$display_row == myrow, ]
-    row_chars[ dcr$display_col ] <- dcr$contents
-    while (any(ncharw(row_chars) > 0)) {
-      for (i in 1:ncol(ht)) {
-        chunk <- substring(row_chars[i], 1, cw[i])
-        if ( (extra <- cw[i] - ncharw(chunk)) > 0) {
-          chunk <- switch(align[i],
-            left = paste0(chunk, str_rep(' ', extra)),
-            center = paste0(str_rep(' ', floor(extra / 2)), chunk, str_rep(' ', ceiling(extra / 2))),
-            right = paste0(str_rep(' ', extra), chunk)
-          )
-        }
-        result <- paste0(result, chunk)
-        row_chars[i] <- substring(row_chars[i], cw[i] + 1)
-        result <- paste0(result, ' ')
-      }
-      result <- paste0(result, '\n')
-    }
-    if (myrow == 1) {
-      dash_row <- paste(str_rep('-', cw), collapse = " ")
-      result <- paste0(result, dash_row, '\n')
-    } else {
-      result <- paste0(result, '\n') # extra blank line for row
-    }
+  cw <- col_width(ht)
+  if (! is.numeric(cw) || anyNA(cw)) cw <- rep(1, ncol(ht))
+  cw <- floor(cw / sum(cw) * (max_width - ncol(ht) + 1))
+
+  charmat_data <- character_matrix(ht, inner_border_h = 1, outer_border_h = 1, inner_border_v = 1, outer_border_v = 1,
+        max_widths = cw)
+  charmat <- charmat_data$charmat
+  border_rows <- charmat_data$border_rows
+  border_cols <- charmat_data$border_cols
+  # if you have a header, you need a whole line of dashes. Otherwise just to indicate columns
+  dash_cols <- if (header) seq_len(ncol(charmat)) else - (border_cols)
+  charmat[c(1, nrow(charmat)), dash_cols] <- '-'
+  if (header) {
+    omit_cols <- border_cols[ - c(1, ncol(ht) + 1)] # skip interior borders
+    charmat[border_rows[2], - omit_cols ] <- '-'
   }
-  result <- paste0(result, str_rep('-', width), '\n')
+
+  result <- paste((apply(charmat, 1, paste0, collapse='')), collapse='\n')
+  result <- paste0(result, '\n\n')
   if (! is.na(cap <- caption(ht))) result <- paste0(result, 'Table: ', cap, '\n')
-  result <- paste0(result, '\n')
+
   result
 }
+
+
+# function to calculate text column widths, wrap huxtable text accordingly, and return a matrix of characters, without
+# borders
+character_matrix <- function (ht, inner_border_h, inner_border_v, outer_border_h, outer_border_v, max_widths = Inf) {
+  dc <- huxtable:::display_cells(ht, all = FALSE)
+  dc <- dc[order(dc$colspan), ]
+  contents <- huxtable:::clean_contents(ht, type = 'screen')
+  drow_mat <- as.matrix(dc[, c('display_row', 'display_col')])
+  dc$contents <- contents[drow_mat]
+
+  widths <- rep(3, ncol(ht))
+  content_widths <- ncharw(dc$contents)
+  max_word_widths <- sapply(lapply(strsplit(dc$contents, "\\s"), ncharw), max)
+  for (r in seq_len(nrow(dc))) {
+    width <- if (wrap(ht)[ dc$display_row[r], dc$display_col[r] ]) max_word_widths[r] else content_widths[r]
+    cols <- seq(dc$display_col[r], dc$end_col[r])
+    # allows for width of interior borders if a cell spans multiple columns
+    if (sum(widths[cols]) + inner_border_h * (dc$colspan[r] - 1) < width) {
+      widths[cols] <- pmax(widths[cols], ceiling(width / dc$colspan[r]))
+    }
+  }
+
+  widths <- pmin(widths, max_widths)
+
+  dc$strings <- vector(nrow(dc), mode = 'list')
+  for (r in seq_len(nrow(dc))) {
+    dcell <- dc[r, ]
+    col <- dcell$display_col
+    end_col <- dcell$end_col
+    width <- sum(widths[col:end_col])
+    strings <- strwrap(dcell$contents, width = width + 1) # for the + 1 see ?strwrap
+    strings <- str_pad(strings, align(ht)[ dcell$display_row, dcell$display_col ], width)
+    dc$strings[[r]] <- strings
+  }
+  dc$text_height <- sapply(dc$strings, length)
+  dc$text_width <- sapply(dc$strings, function (x) max(ncharw(x)))
+
+
+  # row heights as widths: start at 0 and increase it if its too little, sharing equally among relevant cols
+  dc <- dc[order(dc$rowspan),]
+  heights <- rep(1, nrow(ht))
+  for (r in seq_len(nrow(dc))) {
+    dcell <- dc[r, ]
+    rows <- seq(dcell$display_row, dcell$end_row)
+    if (sum(heights[rows]) + inner_border_v * (dcell$rowspan - 1) < dcell$text_height) {
+      heights[rows] <- pmax(heights[rows], ceiling(dcell$text_height / dcell$rowspan))
+    }
+  }
+
+
+  border_widths <- c(outer_border_h, rep(inner_border_h, ncol(ht) - 1), outer_border_h)
+  # width of outer border, then cells + following border:
+  all_widths    <- border_widths + c(0, widths)
+  starting_cols <- cumsum(all_widths[seq_len(ncol(ht))]) + 1
+  border_cols   <- c(starting_cols, sum(all_widths) + 1) - border_widths
+
+  border_heights <- c(outer_border_v, rep(inner_border_v, nrow(ht) - 1), outer_border_v)
+  all_heights    <- border_heights + c(0, heights)
+  starting_rows <- cumsum(all_heights[seq_len(nrow(ht))]) + 1
+  border_rows   <- c(starting_rows, sum(all_heights) + 1) - border_heights
+  charmat <- matrix(' ', sum(all_heights), sum(all_widths))
+
+  for (r in seq_len(nrow(dc))) {
+    dcell <- dc[r, ]
+    string_letters <- unlist(strsplit(dcell$strings[[1]], ''))
+    drow <- dcell$display_row
+    dcol <- dcell$display_col
+    rows <- seq(starting_rows[drow], starting_rows[drow] + dcell$text_height - 1)
+    cols <- seq(starting_cols[dcol], starting_cols[dcol] + dcell$text_width - 1)
+    charmat[rows, cols] <- matrix(string_letters, length(rows), length(cols), byrow = TRUE)
+  }
+
+  list(charmat = charmat, border_rows = border_rows, border_cols = border_cols)
+}
+
+
+ncharw <- function (x) nchar(x, type = 'width')
+
+
+str_pad <- function (strings, align, strlen) {
+  stencils <- rep(str_rep(' ', strlen), length(strings))
+
+  if (align == 'left') {
+    start <- 1
+    stop <- ncharw(strings)
+  } else if (align == 'center') {
+    start <- floor( (ncharw(stencils) - ncharw(strings)) / 2) + 1
+    stop  <- ncharw(stencils) - start + 1
+  } else {
+    start <- ncharw(stencils) - ncharw(strings) + 1
+    stop  <- ncharw(stencils)
+  }
+  substr(stencils, start, stop) <- strings
+
+  stencils
+}
+
 
 str_rep <- function(x, times) {
   mapply(function (s, t) paste0(rep(s, t), collapse = ''), x, times)